refactor(SSG): separate middleware logic (#2315)
* move to middleware

* move core logic

* denoify
