+++ src/helper/ssg/ssg.test.tsx
@@ -3,22 +3,14 @@ import { Hono } from '../../hono'
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { jsx } from '../../jsx'
 import { poweredBy } from '../../middleware/powered-by'
-import {
-  SSG_DISABLED_RESPONSE,
-  fetchRoutesContent,
-  saveContentToFile,
-  ssgParams,
-  toSSG,
-  isSSGContext,
-  disableSSG,
-  onlySSG,
-} from './index'
+import { SSG_DISABLED_RESPONSE, ssgParams, isSSGContext, disableSSG, onlySSG } from './middleware'
+import { fetchRoutesContent, saveContentToFile, toSSG } from './ssg'
 import type {
   BeforeRequestHook,
   AfterResponseHook,
   AfterGenerateHook,
   FileSystemModule,
-} from './index'
+} from './ssg'
 
 const resolveRoutesContent = async (res: ReturnType<typeof fetchRoutesContent>) => {
   const htmlMap = new Map<string, { content: string | ArrayBuffer; mimeType: string }>()
